Provided support for fixing MOBSDK-518 & MOBSDK-519 via the ability to checkin and checkout content, thus forcing a version to be created. Tests added to validate the new functionality but more to come.
Also pulled in the latest CMIS library.

+++ AlfrescoSDK/AlfrescoSDKTests/AlfrescoWorkflowProcessTests.m
@@ -568,7 +568,7 @@ static NSString * const kAlfrescoJBPMAdhocProcessDefinition = @"jbpm$wf:adhoc";
                     else
                     {
                         XCTAssertNotNil(attachmentNodes, @"array should not be nil");
-                        XCTAssertTrue(attachmentNodes.count == attachmentArray.count, @"Array should contain %d attachment(s)", attachmentArray.count);
+                        XCTAssertTrue(attachmentNodes.count == attachmentArray.count, @"Array should contain %lu attachment(s)", (unsigned long)attachmentArray.count);
                         
                         [self deleteCreatedTestProcess:createdProcess completionBlock:^(BOOL succeeded, NSError *deleteError) {
                             XCTAssertTrue(succeeded, @"Deletion flag should be true");
